+++ src/main/java/econo/app/sleeper/domain/user/UserRepository.java
@@ -1,4 +1,4 @@
-package econo.app.sleeper.repository;
+package econo.app.sleeper.domain.user;
 
 import econo.app.sleeper.domain.diary.Diary;
 import econo.app.sleeper.domain.user.User;
