+++ src/main/java/econo/app/sleeper/service/calendar/CalendarService.java
@@ -5,8 +5,8 @@ import econo.app.sleeper.domain.diary.Diary;
 import econo.app.sleeper.domain.sleep.Sleep;
 import econo.app.sleeper.exception.RestApiException;
 import econo.app.sleeper.exception.error.CommonErrorCode;
-import econo.app.sleeper.repository.DiaryRepository;
-import econo.app.sleeper.repository.SleepRepository;
+import econo.app.sleeper.domain.diary.DiaryRepository;
+import econo.app.sleeper.domain.sleep.SleepRepository;
 import econo.app.sleeper.web.calendar.CalendarDateDto;
 import econo.app.sleeper.web.calendar.CalendarDto;
 import econo.app.sleeper.web.calendar.CalendarRequest;
