+++ src/main/java/econo/app/sleeper/service/diary/DiaryService.java
@@ -5,8 +5,8 @@ import econo.app.sleeper.domain.diary.Diary;
 import econo.app.sleeper.domain.user.User;
 import econo.app.sleeper.exception.RestApiException;
 import econo.app.sleeper.exception.error.CommonErrorCode;
-import econo.app.sleeper.repository.DiaryRepository;
-import econo.app.sleeper.repository.UserRepository;
+import econo.app.sleeper.domain.diary.DiaryRepository;
+import econo.app.sleeper.domain.user.UserRepository;
 import econo.app.sleeper.web.diary.dto.DiaryFindDto;
 import econo.app.sleeper.web.diary.dto.DiaryRequest;
 import lombok.RequiredArgsConstructor;
