+++ src/main/java/econo/app/sleeper/service/sleep/SleepAdvisorService.java
@@ -3,12 +3,11 @@ package econo.app.sleeper.service.sleep;
 import econo.app.sleeper.domain.sleep.SleepAdvisor;
 import econo.app.sleeper.exception.RestApiException;
 import econo.app.sleeper.exception.error.CommonErrorCode;
-import econo.app.sleeper.repository.SleepAdvisorRepository;
+import econo.app.sleeper.domain.sleep.SleepAdvisorRepository;
 import econo.app.sleeper.web.sleep.dto.InitialSleepAdvisorDto;
 import econo.app.sleeper.web.sleep.dto.RecommendedTimes;
 import econo.app.sleeper.web.sleep.dto.SleepAdvisorDto;
 import econo.app.sleeper.web.sleep.dto.SleepAdvisorResponse;
-import econo.app.sleeper.web.user.dto.SignUpRequest;
 import econo.app.sleeper.web.user.dto.WakeTimeRecommendDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
