+++ src/main/java/econo/app/sleeper/service/sleep/SleepService.java
@@ -5,8 +5,8 @@ import econo.app.sleeper.domain.sleep.Sleep;
 import econo.app.sleeper.domain.user.User;
 import econo.app.sleeper.exception.RestApiException;
 import econo.app.sleeper.exception.error.CommonErrorCode;
-import econo.app.sleeper.repository.SleepRepository;
-import econo.app.sleeper.repository.UserRepository;
+import econo.app.sleeper.domain.sleep.SleepRepository;
+import econo.app.sleeper.domain.user.UserRepository;
 import econo.app.sleeper.web.sleep.dto.SleepDto;
 import econo.app.sleeper.web.sleep.dto.SleepRequest;
 import lombok.RequiredArgsConstructor;
