+++ src/test/java/econo/app/sleeper/repository/CharacterRepositoryTest.java
@@ -1,7 +1,8 @@
 package econo.app.sleeper.repository;
 
 import econo.app.sleeper.domain.character.Character;
-import org.assertj.core.api.Assertions;
+import econo.app.sleeper.domain.character.CharacterRepository;
+import econo.app.sleeper.domain.user.UserRepository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
