+++ src/test/java/econo/app/sleeper/repository/DiaryRepositoryTest.java
@@ -1,7 +1,9 @@
 package econo.app.sleeper.repository;
 
 import econo.app.sleeper.domain.diary.Diary;
+import econo.app.sleeper.domain.diary.DiaryRepository;
 import econo.app.sleeper.domain.user.User;
+import econo.app.sleeper.domain.user.UserRepository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
