+++ src/test/java/econo/app/sleeper/repository/UserRepositoryTest.java
@@ -1,6 +1,7 @@
 package econo.app.sleeper.repository;
 
 import econo.app.sleeper.domain.user.User;
+import econo.app.sleeper.domain.user.UserRepository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
