+++ src/test/java/econo/app/sleeper/service/sleep/SleepServiceTest.java
@@ -1,18 +1,13 @@
 package econo.app.sleeper.service.sleep;
 
-import econo.app.sleeper.domain.sleep.Sleep;
-import econo.app.sleeper.repository.CharacterRepository;
-import econo.app.sleeper.repository.SleepRepository;
-import org.assertj.core.api.Assertions;
-import org.junit.Test;
+import econo.app.sleeper.domain.character.CharacterRepository;
+import econo.app.sleeper.domain.sleep.SleepRepository;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.time.*;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
